- Remove invalid `related_object` usage from report tasks - Guard report job state updates when job lookup fails - Use settings-based report cache TTL instead of model attribute - refactored metric tasks into task module for easier sorting -removed audit log entries fro department dashbaord data
