+++ inventory_metrics/tasks/__init__.py
@@ -1,0 +1,2 @@
+from .reports import *
+from .snapshots import *
